// packages/measurements/src/measure.types.ts
interface HorizontalMeasurementDrawing extends PolylineDrawing {
  distance: number
  distances: Array<number>
  labels: Array<AnyObject>
  angles: Array<number>
  dashedLines?: Array<AnyObject>
  heightPlane?: Cesium.Plane
  heightPlaneCV?: Cesium.Plane
  height?: number
  firstMove?: boolean
  tempNextPos: Cesium.Cartesian3
}

interface VerticalMeasurementPolylineSegment extends PolylineSegment {
  distance?: number
  labelPosition?: Cesium.Cartesian3
  draggingPlane: Cesium.Plane
  surfaceNormal: Cesium.Cartesian3
}

interface HeightMeasurementPolylineSegment extends PolylineSegment {
  distance: number
  labelPosition: Cesium.Cartesian3
}

interface PolylineDrawing {
  show: boolean
  positions: Array<Cesium.Cartesian3>
  tempPositions: Array<Cesium.Cartesian3>
  drawStatus: number
  loop?: boolean
}